fixing e2e import, moving sentry state into seperate file
